+++ tests/TestRun2/Config/stockIds.csv
@@ -1,2 +1,2 @@
 googleStockId,appleStockId
-3bc9d3a8-d3e3-47e4-8e3c-1c8bb76a2535,b1dbf030-ad75-4f8b-a0e1-b8f4d2b19eee
+1448f7bd-d0e3-4daf-ae08-e243bf6b036d,01184418-d5c2-4ddd-8105-b7a4ad7873fd
